# [WIP] storage: clone Spans before adding them to tscache.Request and use sync.Pool