Related to #16796.

This PR makes two related changes:
- it clones all Spans before adding them to a `tscache.Request`
- it adds `*tscache.Request` to a `sync.Pool`

Both of these were done to reduce GC time. Using GCTrace on my linux machine (since
it seems to be broken on Darwin) a read-only workload originally showed a consistent 
**7%** of time spent in GC. This was when running:
```
kv --read-percent 100 --concurrency 16 --splits 100 --write-seq 100000
```

After the first change, this time dropped to **3%**. After the second change, it dropped
again to **2%**. 

The second change is pretty straightforward. The first is more interesting, especially in
light of the experimentation done in #16796. If I individually deep copy the `Spans` by
adding a `Clone` method to `roachpb.Span`, I don't see the same effect as I do here. I
had originally thought that the old GC time was in-part due to scanning the corresponding
`BatchRequest` and `endCmds` structs, which were for some reason being referenced and
therefore unable to be GCed. This proves that that was not the case. This leads me to believe
that the GC cost reduction is due to the single slab of memory being allocated for all keys in
a `BatchRequest`. This theory is fairly unsubstantiated though. I'm curious if anyone has any
other ideas as to why we're seeing such a big difference.

In addition to the GC impact, the two commits also has a fairly large impact on `kv`
performance as a whole. The following are the results from three 1 minute `kv` runs
before and after the changes. As would be expected from a GC related change like this,
we see the impact on the tail latencies. Specifically, it reduces 99th percentile latency by
about **33%**. It also increases throughput by about **7%**.

### Before
```
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
   60.0s        0        1985392        33089.4      0.5      0.4      0.9      2.6     30.4

BenchmarkLoadgenKV/readPercent=100/splits=100/concurrency=16/duration=1m0s	 1985395	     30221 ns/op

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
   60.0s        0        1979390        32989.3      0.5      0.4      0.9      2.8     27.3

BenchmarkLoadgenKV/readPercent=100/splits=100/concurrency=16/duration=1m0s	 1979390	     30312 ns/op

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
   60.0s        0        1986851        33113.6      0.5      0.4      0.9      2.8     27.3

BenchmarkLoadgenKV/readPercent=100/splits=100/concurrency=16/duration=1m0s	 1986852	     30199 ns/op
```

### After
```
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
   60.0s        0        2121080        35350.8      0.5      0.4      0.9      1.8     23.1

BenchmarkLoadgenKV/readPercent=100/splits=100/concurrency=16/duration=1m0s	 2121080	     28287 ns/op

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
   60.0s        0        2128153        35468.7      0.5      0.4      0.8      1.8     27.3

BenchmarkLoadgenKV/readPercent=100/splits=100/concurrency=16/duration=1m0s	 2128153	     28193 ns/op

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
   60.0s        0        2123837        35396.8      0.5      0.4      0.8      1.8     26.2

BenchmarkLoadgenKV/readPercent=100/splits=100/concurrency=16/duration=1m0s	 2123847	     28251 ns/op
```