@tschottdorf 2017-10-20
Nice! <img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/> mod TODOs and @petermattis.

---

Reviewed 3 of 3 files at r1, 6 of 6 files at r2.
Review status: all files reviewed at latest revision, 4 unresolved discussions, all commit checks successful.

---

*[pkg/storage/replica.go, line 1941 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/19418#-KwvnEkIpZTCgUpPGsSJ:-KwvnEkIpZTCgUpPGsSK:bq4l77b) ([raw file](https://github.com/cockroachdb/cockroach/blob/9a95e3239a7b41a36aad2ce14583efc8daa383f9/pkg/storage/replica.go#L1941)):*
> ```Go
> 	}
> 
> 	var alloc bufalloc.ByteAllocator
> ```

Add commentary explaining why copying spans is the better approach.

---

*[pkg/storage/replica.go, line 1989 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/19418#-KwvndJXGRA3hE7YypHe:-KwvndJXGRA3hE7YypHf:b-tgi9tp) ([raw file](https://github.com/cockroachdb/cockroach/blob/e50a14fb4fbd0e50319a9e7ce0243334d4ac1eb7/pkg/storage/replica.go#L1989)):*
> ```Go
> 					// the span.
> 
> 					// Make a copy of the key to avoid holding on to the large buffer the
> ```

This was a bad idea then or just no measurable difference?

---

*[pkg/storage/tscache/cache_impl.go, line 573 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/19418#-Kwvp5AmF5ehqUhS2929:-Kwvp5AmF5ehqUhS292A:bs3wdpu) ([raw file](https://github.com/cockroachdb/cockroach/blob/68c2e547b457a3ae3a1797d190c675c83000c19d/pkg/storage/tscache/cache_impl.go#L573)):*
> ```Go
> 	// Expand the requests, inserting the spans into either the read or write
> 	// interval caches.
> 	for _, req := range reqs {
> ```

Does it make sense to zero out `reqs` while you're at it? The thinking being that `reqs` would otherwise continue to reference `req` and GC has to look at that.

```go
for i := range reqs {
  req, reqs[i] = reqs[i], nil
  // ...
}
```

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/19418#-:-Kwvppm_m3DQeVoT-qdX:bhrc17b)*
<!-- Sent from Reviewable.io -->