@petermattis 2017-10-20
That's a nice win.

---

Review status: all files reviewed at latest revision, 5 unresolved discussions, all commit checks successful.

---

*[pkg/storage/replica.go, line 1941 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/19418#-KwvnEkIpZTCgUpPGsSJ:-Kww5--gLReqQ8Oh5cEC:b-tdgutz) ([raw file](https://github.com/cockroachdb/cockroach/blob/9a95e3239a7b41a36aad2ce14583efc8daa383f9/pkg/storage/replica.go#L1941)):*
<details><summary><i>Previously, tschottdorf (Tobias Schottdorf) wrote…</i></summary><blockquote>

Add commentary explaining why copying spans is the better approach.
</blockquote></details>

There is a serious downside to this approach: the minimum chunk size used by `bufalloc.ByteAllocator` is 512 bytes. 

As an alternative, perhaps we should make a first pass over the request gathering the spans and their total size and then allocate a buffer of exactly that size.

---

*[pkg/storage/replica.go, line 1953 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/19418#-Kww4jfDzgLImeDxRvLc:-Kww4jfDzgLImeDxRvLd:begmxm) ([raw file](https://github.com/cockroachdb/cockroach/blob/9a95e3239a7b41a36aad2ce14583efc8daa383f9/pkg/storage/replica.go#L1953)):*
> ```Go
> 				// DeleteRange adds to the write timestamp cache to prevent
> 				// subsequent writes from rewriting history.
> 				alloc, headerCopy = alloc.CopySpan(header)
> ```

Was this actually necessary? Keys that come from the request should have been allocated separately. Or perhaps bundling them into big allocation buffers is a win for the GC. Curious to understand this better.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/19418#-:-Kww5YB8QFwvmdVtGiqm:bqa7z7y)*
<!-- Sent from Reviewable.io -->